+++ app/models/version/billing/price_version.rb
@@ -1,0 +1,7 @@
+module Billing
+  class PriceVersion < PaperTrail::Version
+    self.table_name    = :log_prices
+    self.sequence_name = :log_prices_id_seq
+  end
+end
+