+++ db/migrate/20200910085157_change_invoice_item_price_scale_to_three_places.rb
@@ -1,0 +1,5 @@
+class ChangeInvoiceItemPriceScaleToThreePlaces < ActiveRecord::Migration[6.0]
+  def change
+    change_column :invoice_items, :price, :decimal, precision: 10, scale: 3
+  end
+end