+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+15.09.2020
+* Fixed e-invoice sending issue with QUE [#1683](https://github.com/internetee/registry/pull/1683)
+
+14.09.2020
+* Restored version logging for registry prices [#980](https://github.com/internetee/registry/pull/980)
+
+11.09.2020
+* Registrars can now top up their credit accounts without generating invoice in advance [#1101](https://github.com/internetee/registry/issues/1101)
+* Fixed typo in admin settings [#371](https://github.com/internetee/registry/issues/371)
+
+10.09.2020
+* New registrar ref nr are now always created 7 digits long [#1679](https://github.com/internetee/registry/pull/1679)
+
+08.09.2020
+* Removed bank statement import option [#1674](https://github.com/internetee/registry/pull/1674)
+* Fixed error with reference nr not being found in the transaction [#1677](https://github.com/internetee/registry/issues/1677)
+
+04.09.2020
+* Removed reduntant domains.registered_at db column [#1445](https://github.com/internetee/registry/pull/1445)
+* Certificate revocation lists are now hanlded outside of the application code [#1662](https://github.com/internetee/registry/pull/1662)
+* Monthly invoices are sent one by one to elliminate reply delay from accounting system [#1671](https://github.com/internetee/registry/pull/1671)
+* Fixed poll request ip whitelist issue [#1672](https://github.com/internetee/registry/pull/1672)
+
 03.09.2020
 * Refactored session timeout management [#711](https://github.com/internetee/registry/issues/711)
 * Improved error handling for epp requests without proper session [#1276](https://github.com/internetee/registry/pull/1276)